requirements.txt
# chcp 65001 -> 處理cp950 error
# 核心框架 (CoreFrameworks)
# Successfully installed langchain-1.2.15 langchain-classic-1.0.4 langchain-community-0.4.1 langchain-core-1.3.2 langchain-protocol-0.0.12
langchain=0.3.7
langchain-core=0.1.0
langchain-community=0.3.7
langchain-openai>=0.0.5
langgraph>=0.0.26

# 文件解析與處理 (Parsing&Processing)
docling>=1.0.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0

# 檢索與向量資料庫 (Retrieval&VectorDB)
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.1

# 工具與其他 (Utilities)
pydantic>=2.0.0
pathlib>=1.0.1
tqdm>=4.66.0
tiktoken>=0.5.0